Update goals
The SCC content team did a review of the goals that are available. This PR reflects those changes. I also found that some of the goals in the list did not have a corresponding goal in SCC, so I removed those. Please let me know if you've got any questions! :)