Update cos-events-at.md
This file was correct in the f1881/feature-cos... branch until I ran a merge to draft the morning of 052124, and after that 47 updates were pushed to the feature branch and this file reverted to an earlier copy: 041724.  The last revision date that I recall was 050224 or 050624.  This is not correct.  See #1767 for earlier work.